+++ x/collection/keeper/keeper_test.go
@@ -141,8 +141,8 @@ func (s *KeeperTestSuite) SetupTest() {
 		// create a chain of its length depth_limit
 		for i := range tokens[1:] {
 			r := len(tokens) - 1 - i
-			subject := tokens[r].Id
-			target := tokens[r-1].Id
+			subject := tokens[r].TokenId
+			target := tokens[r-1].TokenId
 			err := s.keeper.Attach(s.ctx, s.contractID, to, subject, target)
 			s.Require().NoError(err)
 		}
@@ -151,7 +151,7 @@ func (s *KeeperTestSuite) SetupTest() {
 		s.Require().NoError(err)
 
 		// a chain of length 2
-		err = s.keeper.Attach(s.ctx, s.contractID, to, tokens[remainders-1].Id, tokens[remainders-2].Id)
+		err = s.keeper.Attach(s.ctx, s.contractID, to, tokens[remainders-1].TokenId, tokens[remainders-2].TokenId)
 		s.Require().NoError(err)
 
 	}