+++ x/collection/keeper/msg_server_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/line/lbm-sdk/x/token/class"
 )
 
-func (s *KeeperTestSuite) TestMsgTransferFT() {
+func (s *KeeperTestSuite) TestMsgSendFT() {
 	testCases := map[string]struct {
 		contractID string
 		amount     sdk.Int
@@ -32,7 +32,7 @@ func (s *KeeperTestSuite) TestMsgTransferFT() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgTransferFT{
+			req := &collection.MsgSendFT{
 				ContractId: tc.contractID,
 				From:       s.vendor.String(),
 				To:         s.customer.String(),
@@ -40,7 +40,7 @@ func (s *KeeperTestSuite) TestMsgTransferFT() {
 					collection.NewFTCoin(s.ftClassID, tc.amount),
 				),
 			}
-			res, err := s.msgServer.TransferFT(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.SendFT(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -51,37 +51,37 @@ func (s *KeeperTestSuite) TestMsgTransferFT() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgTransferFTFrom() {
+func (s *KeeperTestSuite) TestMsgOperatorSendFT() {
 	testCases := map[string]struct {
 		contractID string
-		proxy      sdk.AccAddress
+		operator   sdk.AccAddress
 		from       sdk.AccAddress
 		amount     sdk.Int
 		err        error
 	}{
 		"valid request": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			amount:     s.balance,
 		},
 		"contract not found": {
 			contractID: "deadbeef",
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			amount:     s.balance,
 			err:        class.ErrContractNotExist,
 		},
 		"not approved": {
 			contractID: s.contractID,
-			proxy:      s.vendor,
+			operator:   s.vendor,
 			from:       s.customer,
 			amount:     s.balance,
 			err:        collection.ErrCollectionNotApproved,
 		},
 		"insufficient funds": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			amount:     s.balance.Add(sdk.OneInt()),
 			err:        collection.ErrInsufficientToken,
@@ -92,16 +92,16 @@ func (s *KeeperTestSuite) TestMsgTransferFTFrom() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgTransferFTFrom{
+			req := &collection.MsgOperatorSendFT{
 				ContractId: tc.contractID,
-				Proxy:      tc.proxy.String(),
+				Operator:   tc.operator.String(),
 				From:       tc.from.String(),
 				To:         s.vendor.String(),
 				Amount: collection.NewCoins(
 					collection.NewFTCoin(s.ftClassID, tc.amount),
 				),
 			}
-			res, err := s.msgServer.TransferFTFrom(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.OperatorSendFT(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -112,7 +112,7 @@ func (s *KeeperTestSuite) TestMsgTransferFTFrom() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgTransferNFT() {
+func (s *KeeperTestSuite) TestMsgSendNFT() {
 	testCases := map[string]struct {
 		contractID string
 		tokenID    string
@@ -148,13 +148,13 @@ func (s *KeeperTestSuite) TestMsgTransferNFT() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgTransferNFT{
+			req := &collection.MsgSendNFT{
 				ContractId: tc.contractID,
 				From:       s.customer.String(),
 				To:         s.vendor.String(),
 				TokenIds:   []string{tc.tokenID},
 			}
-			res, err := s.msgServer.TransferNFT(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.SendNFT(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -165,52 +165,52 @@ func (s *KeeperTestSuite) TestMsgTransferNFT() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgTransferNFTFrom() {
+func (s *KeeperTestSuite) TestMsgOperatorSendNFT() {
 	tokenID := collection.NewNFTID(s.nftClassID, 1)
 	testCases := map[string]struct {
 		contractID string
-		proxy      sdk.AccAddress
+		operator   sdk.AccAddress
 		from       sdk.AccAddress
 		tokenID    string
 		err        error
 	}{
 		"valid request": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenID:    tokenID,
 		},
 		"contract not found": {
 			contractID: "deadbeef",
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenID:    tokenID,
 			err:        class.ErrContractNotExist,
 		},
 		"not approved": {
 			contractID: s.contractID,
-			proxy:      s.vendor,
+			operator:   s.vendor,
 			from:       s.customer,
 			tokenID:    tokenID,
 			err:        collection.ErrCollectionNotApproved,
 		},
 		"not found": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenID:    collection.NewNFTID("deadbeef", 1),
 			err:        collection.ErrTokenNotExist,
 		},
 		"child": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenID:    collection.NewNFTID(s.nftClassID, 2),
 			err:        collection.ErrTokenCannotTransferChildToken,
 		},
 		"not owned by": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenID:    collection.NewNFTID(s.nftClassID, s.numNFTs+1),
 			err:        collection.ErrTokenNotOwnedBy,
@@ -221,14 +221,14 @@ func (s *KeeperTestSuite) TestMsgTransferNFTFrom() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgTransferNFTFrom{
+			req := &collection.MsgOperatorSendNFT{
 				ContractId: tc.contractID,
-				Proxy:      tc.proxy.String(),
+				Operator:   tc.operator.String(),
 				From:       tc.from.String(),
 				To:         s.vendor.String(),
 				TokenIds:   []string{tc.tokenID},
 			}
-			res, err := s.msgServer.TransferNFTFrom(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.OperatorSendNFT(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -239,28 +239,28 @@ func (s *KeeperTestSuite) TestMsgTransferNFTFrom() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgApprove() {
+func (s *KeeperTestSuite) TestMsgAuthorizeOperator() {
 	testCases := map[string]struct {
 		contractID string
-		approver   sdk.AccAddress
-		proxy      sdk.AccAddress
+		holder     sdk.AccAddress
+		operator   sdk.AccAddress
 		err        error
 	}{
 		"valid request": {
 			contractID: s.contractID,
-			approver:   s.customer,
-			proxy:      s.vendor,
+			holder:     s.customer,
+			operator:   s.vendor,
 		},
 		"contract not found": {
 			contractID: "deadbeef",
-			approver:   s.customer,
-			proxy:      s.vendor,
+			holder:     s.customer,
+			operator:   s.vendor,
 			err:        class.ErrContractNotExist,
 		},
 		"already approved": {
 			contractID: s.contractID,
-			approver:   s.customer,
-			proxy:      s.operator,
+			holder:     s.customer,
+			operator:   s.operator,
 			err:        collection.ErrCollectionAlreadyApproved,
 		},
 	}
@@ -269,12 +269,12 @@ func (s *KeeperTestSuite) TestMsgApprove() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgApprove{
+			req := &collection.MsgAuthorizeOperator{
 				ContractId: tc.contractID,
-				Approver:   tc.approver.String(),
-				Proxy:      tc.proxy.String(),
+				Holder:     tc.holder.String(),
+				Operator:   tc.operator.String(),
 			}
-			res, err := s.msgServer.Approve(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.AuthorizeOperator(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -285,28 +285,28 @@ func (s *KeeperTestSuite) TestMsgApprove() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgDisapprove() {
+func (s *KeeperTestSuite) TestMsgRevokeOperator() {
 	testCases := map[string]struct {
 		contractID string
-		approver   sdk.AccAddress
-		proxy      sdk.AccAddress
+		holder     sdk.AccAddress
+		operator   sdk.AccAddress
 		err        error
 	}{
 		"valid request": {
 			contractID: s.contractID,
-			approver:   s.customer,
-			proxy:      s.operator,
+			holder:     s.customer,
+			operator:   s.operator,
 		},
 		"contract not found": {
 			contractID: "deadbeef",
-			approver:   s.customer,
-			proxy:      s.operator,
+			holder:     s.customer,
+			operator:   s.operator,
 			err:        class.ErrContractNotExist,
 		},
 		"no authorization": {
 			contractID: s.contractID,
-			approver:   s.customer,
-			proxy:      s.vendor,
+			holder:     s.customer,
+			operator:   s.vendor,
 			err:        collection.ErrCollectionNotApproved,
 		},
 	}
@@ -315,12 +315,12 @@ func (s *KeeperTestSuite) TestMsgDisapprove() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgDisapprove{
+			req := &collection.MsgRevokeOperator{
 				ContractId: tc.contractID,
-				Approver:   tc.approver.String(),
-				Proxy:      tc.proxy.String(),
+				Holder:     tc.holder.String(),
+				Operator:   tc.operator.String(),
 			}
-			res, err := s.msgServer.Disapprove(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.RevokeOperator(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -625,47 +625,47 @@ func (s *KeeperTestSuite) TestMsgBurnFT() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgBurnFTFrom() {
+func (s *KeeperTestSuite) TestMsgOperatorBurnFT() {
 	amount := collection.NewCoins(
 		collection.NewFTCoin(s.ftClassID, s.balance),
 	)
 	testCases := map[string]struct {
 		contractID string
-		proxy      sdk.AccAddress
+		operator   sdk.AccAddress
 		from       sdk.AccAddress
 		amount     []collection.Coin
 		err        error
 	}{
 		"valid request": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			amount:     amount,
 		},
 		"contract not found": {
 			contractID: "deadbeef",
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			amount:     amount,
 			err:        class.ErrContractNotExist,
 		},
 		"no authorization": {
 			contractID: s.contractID,
-			proxy:      s.vendor,
+			operator:   s.vendor,
 			from:       s.customer,
 			amount:     amount,
 			err:        collection.ErrCollectionNotApproved,
 		},
 		"no permission": {
 			contractID: s.contractID,
-			proxy:      s.stranger,
+			operator:   s.stranger,
 			from:       s.customer,
 			amount:     amount,
 			err:        collection.ErrTokenNoPermission,
 		},
 		"insufficient funds": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			amount: collection.NewCoins(
 				collection.NewFTCoin("00bab10c", sdk.OneInt()),
@@ -678,13 +678,13 @@ func (s *KeeperTestSuite) TestMsgBurnFTFrom() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgBurnFTFrom{
+			req := &collection.MsgOperatorBurnFT{
 				ContractId: tc.contractID,
-				Proxy:      tc.proxy.String(),
+				Operator:   tc.operator.String(),
 				From:       tc.from.String(),
 				Amount:     tc.amount,
 			}
-			res, err := s.msgServer.BurnFTFrom(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.OperatorBurnFT(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -768,47 +768,47 @@ func (s *KeeperTestSuite) TestMsgBurnNFT() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgBurnNFTFrom() {
+func (s *KeeperTestSuite) TestMsgOperatorBurnNFT() {
 	tokenIDs := []string{
 		collection.NewNFTID(s.nftClassID, 1),
 	}
 	testCases := map[string]struct {
 		contractID string
-		proxy      sdk.AccAddress
+		operator   sdk.AccAddress
 		from       sdk.AccAddress
 		tokenIDs   []string
 		err        error
 	}{
 		"valid request": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenIDs:   tokenIDs,
 		},
 		"contract not found": {
 			contractID: "deadbeef",
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenIDs:   tokenIDs,
 			err:        class.ErrContractNotExist,
 		},
 		"no authorization": {
 			contractID: s.contractID,
-			proxy:      s.vendor,
+			operator:   s.vendor,
 			from:       s.customer,
 			tokenIDs:   tokenIDs,
 			err:        collection.ErrCollectionNotApproved,
 		},
 		"no permission": {
 			contractID: s.contractID,
-			proxy:      s.stranger,
+			operator:   s.stranger,
 			from:       s.customer,
 			tokenIDs:   tokenIDs,
 			err:        collection.ErrTokenNoPermission,
 		},
 		"not found": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenIDs: []string{
 				collection.NewNFTID("deadbeef", 1),
@@ -817,7 +817,7 @@ func (s *KeeperTestSuite) TestMsgBurnNFTFrom() {
 		},
 		"child": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenIDs: []string{
 				collection.NewNFTID(s.nftClassID, 2),
@@ -826,7 +826,7 @@ func (s *KeeperTestSuite) TestMsgBurnNFTFrom() {
 		},
 		"not owned by": {
 			contractID: s.contractID,
-			proxy:      s.operator,
+			operator:   s.operator,
 			from:       s.customer,
 			tokenIDs: []string{
 				collection.NewNFTID(s.nftClassID, s.numNFTs+1),
@@ -839,13 +839,13 @@ func (s *KeeperTestSuite) TestMsgBurnNFTFrom() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgBurnNFTFrom{
+			req := &collection.MsgOperatorBurnNFT{
 				ContractId: tc.contractID,
-				Proxy:      tc.proxy.String(),
+				Operator:   tc.operator.String(),
 				From:       tc.from.String(),
 				TokenIds:   tc.tokenIDs,
 			}
-			res, err := s.msgServer.BurnNFTFrom(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.OperatorBurnNFT(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -907,8 +907,8 @@ func (s *KeeperTestSuite) TestMsgModify() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			changes := []collection.Change{{
-				Field: collection.AttributeKeyName.String(),
+			changes := []collection.Attribute{{
+				Key:   collection.AttributeKeyName.String(),
 				Value: "test",
 			}}
 			req := &collection.MsgModify{
@@ -1120,7 +1120,7 @@ func (s *KeeperTestSuite) TestMsgDetach() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgAttachFrom() {
+func (s *KeeperTestSuite) TestMsgOperatorAttach() {
 	testCases := map[string]struct {
 		contractID string
 		operator   sdk.AccAddress
@@ -1161,14 +1161,14 @@ func (s *KeeperTestSuite) TestMsgAttachFrom() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgAttachFrom{
+			req := &collection.MsgOperatorAttach{
 				ContractId: tc.contractID,
-				Proxy:      tc.operator.String(),
+				Operator:   tc.operator.String(),
 				From:       s.customer.String(),
 				TokenId:    tc.subjectID,
 				ToTokenId:  tc.targetID,
 			}
-			res, err := s.msgServer.AttachFrom(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.OperatorAttach(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return
@@ -1179,7 +1179,7 @@ func (s *KeeperTestSuite) TestMsgAttachFrom() {
 	}
 }
 
-func (s *KeeperTestSuite) TestMsgDetachFrom() {
+func (s *KeeperTestSuite) TestMsgOperatorDetach() {
 	testCases := map[string]struct {
 		contractID string
 		operator   sdk.AccAddress
@@ -1215,13 +1215,13 @@ func (s *KeeperTestSuite) TestMsgDetachFrom() {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
 
-			req := &collection.MsgDetachFrom{
+			req := &collection.MsgOperatorDetach{
 				ContractId: tc.contractID,
-				Proxy:      tc.operator.String(),
+				Operator:   tc.operator.String(),
 				From:       s.customer.String(),
 				TokenId:    tc.subjectID,
 			}
-			res, err := s.msgServer.DetachFrom(sdk.WrapSDKContext(ctx), req)
+			res, err := s.msgServer.OperatorDetach(sdk.WrapSDKContext(ctx), req)
 			s.Require().ErrorIs(err, tc.err)
 			if tc.err != nil {
 				return