+++ x/token/client/testutil/query.go
@@ -27,7 +27,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdBalance() {
 	}{
 		"valid query": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.customer.String(),
 			},
 			true,
@@ -37,7 +37,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdBalance() {
 		},
 		"extra args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.customer.String(),
 				"extra",
 			},
@@ -46,14 +46,14 @@ func (s *IntegrationTestSuite) TestNewQueryCmdBalance() {
 		},
 		"not enough args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 			},
 			false,
 			nil,
 		},
 		"invalid address": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				"invalid",
 			},
 			false,
@@ -94,16 +94,16 @@ func (s *IntegrationTestSuite) TestNewQueryCmdToken() {
 	}{
 		"valid query": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 			},
 			true,
-			&token.QueryTokenClassResponse{
-				Class: s.classes[0],
+			&token.QueryContractResponse{
+				Contract: s.classes[0],
 			},
 		},
 		"extra args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				"extra",
 			},
 			false,
@@ -120,7 +120,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdToken() {
 		tc := tc
 
 		s.Run(name, func() {
-			cmd := cli.NewQueryCmdTokenClass()
+			cmd := cli.NewQueryCmdContract()
 			out, err := clitestutil.ExecTestCLICmd(val.ClientCtx, cmd, append(tc.args, commonArgs...))
 			if !tc.valid {
 				s.Require().Error(err)
@@ -128,7 +128,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdToken() {
 			}
 			s.Require().NoError(err)
 
-			var actual token.QueryTokenClassResponse
+			var actual token.QueryContractResponse
 			s.Require().NoError(val.ClientCtx.Codec.UnmarshalJSON(out.Bytes(), &actual), out.String())
 			s.Require().Equal(tc.expected, &actual)
 		})
@@ -150,8 +150,8 @@ func (s *IntegrationTestSuite) TestNewQueryCmdTokens() {
 		"query all": {
 			[]string{},
 			true,
-			&token.QueryTokenClassesResponse{
-				Classes:    s.classes,
+			&token.QueryContractsResponse{
+				Contracts:  s.classes,
 				Pagination: &query.PageResponse{},
 			},
 		},
@@ -168,7 +168,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdTokens() {
 		tc := tc
 
 		s.Run(name, func() {
-			cmd := cli.NewQueryCmdTokenClasses()
+			cmd := cli.NewQueryCmdContracts()
 			out, err := clitestutil.ExecTestCLICmd(val.ClientCtx, cmd, append(tc.args, commonArgs...))
 			if !tc.valid {
 				s.Require().Error(err)
@@ -176,7 +176,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdTokens() {
 			}
 			s.Require().NoError(err)
 
-			var actual token.QueryTokenClassesResponse
+			var actual token.QueryContractsResponse
 			s.Require().NoError(val.ClientCtx.Codec.UnmarshalJSON(out.Bytes(), &actual), out.String())
 			s.Require().Equal(tc.expected, &actual)
 		})
@@ -197,7 +197,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdGranteeGrants() {
 	}{
 		"valid query": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.vendor.String(),
 			},
 			true,
@@ -223,7 +223,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdGranteeGrants() {
 		},
 		"extra args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.vendor.String(),
 				"extra",
 			},
@@ -232,7 +232,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdGranteeGrants() {
 		},
 		"not enough args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 			},
 			false,
 			nil,
@@ -258,7 +258,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdGranteeGrants() {
 	}
 }
 
-func (s *IntegrationTestSuite) TestNewQueryCmdApproved() {
+func (s *IntegrationTestSuite) TestNewQueryCmdIsOperatorFor() {
 	val := s.network.Validators[0]
 	commonArgs := []string{
 		fmt.Sprintf("--%s=%d", flags.FlagHeight, s.setupHeight),
@@ -272,18 +272,18 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApproved() {
 	}{
 		"valid query": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.vendor.String(),
 				s.customer.String(),
 			},
 			true,
-			&token.QueryApprovedResponse{
-				Approved: true,
+			&token.QueryIsOperatorForResponse{
+				Authorized: true,
 			},
 		},
 		"extra args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.vendor.String(),
 				s.customer.String(),
 				"extra",
@@ -293,7 +293,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApproved() {
 		},
 		"not enough args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.vendor.String(),
 			},
 			false,
@@ -305,7 +305,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApproved() {
 		tc := tc
 
 		s.Run(name, func() {
-			cmd := cli.NewQueryCmdApproved()
+			cmd := cli.NewQueryCmdIsOperatorFor()
 			out, err := clitestutil.ExecTestCLICmd(val.ClientCtx, cmd, append(tc.args, commonArgs...))
 			if !tc.valid {
 				s.Require().Error(err)
@@ -313,14 +313,14 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApproved() {
 			}
 			s.Require().NoError(err)
 
-			var actual token.QueryApprovedResponse
+			var actual token.QueryIsOperatorForResponse
 			s.Require().NoError(val.ClientCtx.Codec.UnmarshalJSON(out.Bytes(), &actual), out.String())
 			s.Require().Equal(tc.expected, &actual)
 		})
 	}
 }
 
-func (s *IntegrationTestSuite) TestNewQueryCmdApprovers() {
+func (s *IntegrationTestSuite) TestNewQueryCmdHoldersByOperator() {
 	val := s.network.Validators[0]
 	commonArgs := []string{
 		fmt.Sprintf("--%s=%d", flags.FlagHeight, s.setupHeight),
@@ -334,18 +334,18 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApprovers() {
 	}{
 		"valid query": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.vendor.String(),
 			},
 			true,
-			&token.QueryApproversResponse{
-				Approvers:  []string{s.customer.String()},
+			&token.QueryHoldersByOperatorResponse{
+				Holders:    []string{s.customer.String()},
 				Pagination: &query.PageResponse{},
 			},
 		},
 		"extra args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 				s.vendor.String(),
 				"extra",
 			},
@@ -354,7 +354,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApprovers() {
 		},
 		"not enough args": {
 			[]string{
-				s.classes[0].ContractId,
+				s.classes[0].Id,
 			},
 			false,
 			nil,
@@ -365,7 +365,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApprovers() {
 		tc := tc
 
 		s.Run(name, func() {
-			cmd := cli.NewQueryCmdApprovers()
+			cmd := cli.NewQueryCmdHoldersByOperator()
 			out, err := clitestutil.ExecTestCLICmd(val.ClientCtx, cmd, append(tc.args, commonArgs...))
 			if !tc.valid {
 				s.Require().Error(err)
@@ -373,7 +373,7 @@ func (s *IntegrationTestSuite) TestNewQueryCmdApprovers() {
 			}
 			s.Require().NoError(err)
 
-			var actual token.QueryApproversResponse
+			var actual token.QueryHoldersByOperatorResponse
 			s.Require().NoError(val.ClientCtx.Codec.UnmarshalJSON(out.Bytes(), &actual), out.String())
 			s.Require().Equal(tc.expected, &actual)
 		})