+++ CHANGELOG.md
@@ -48,9 +48,9 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Bug Fixes
 * (client) [\#817](https://github.com/line/lbm-sdk/pull/817) remove support for composite (BLS) type
-
 * (x/foundation) [#834](https://github.com/line/lbm-sdk/pull/834) Apply foundation audit
 * (x/collection,token) [#849](https://github.com/line/lbm-sdk/pull/849) Introduce codespace into x/collection,token
+* (x/token,collection) [#863](https://github.com/line/lbm-sdk/pull/863) Update x/collection,token proto
 
 ### Removed
 * [\#853](https://github.com/line/lbm-sdk/pull/853) remove useless stub BeginBlock, EndBlock methods from modules below